git/git.go: split the code out into multiple files:
* `git/batch_obj_iter.go`
* `git/commit.go`
* `git/git.go`
* `git/git_command.go`
* `git/obj_head_iter.go`
* `git/obj_iter.go`
* `git/oid.go`
* `git/ref_iter.go`
* `git/reference.go`
* `git/tag.go`
* `git/tree.go`
