+++ git/tag.go
@@ -1,0 +1,61 @@
+package git
+
+import (
+	"fmt"
+
+	"github.com/github/git-sizer/counts"
+)
+
+// Tag represents the information that we need about a Git tag object.
+type Tag struct {
+	Size         counts.Count32
+	Referent     OID
+	ReferentType ObjectType
+}
+
+// ParseTag parses the Git tag object whose contents are contained in
+// `data`. `oid` is used only in error messages.
+func ParseTag(oid OID, data []byte) (*Tag, error) {
+	var referent OID
+	var referentFound bool
+	var referentType ObjectType
+	var referentTypeFound bool
+	iter, err := NewObjectHeaderIter(oid.String(), data)
+	if err != nil {
+		return nil, err
+	}
+	for iter.HasNext() {
+		key, value, err := iter.Next()
+		if err != nil {
+			return nil, err
+		}
+		switch key {
+		case "object":
+			if referentFound {
+				return nil, fmt.Errorf("multiple referents found in tag %s", oid)
+			}
+			referent, err = NewOID(value)
+			if err != nil {
+				return nil, fmt.Errorf("malformed object header in tag %s", oid)
+			}
+			referentFound = true
+		case "type":
+			if referentTypeFound {
+				return nil, fmt.Errorf("multiple types found in tag %s", oid)
+			}
+			referentType = ObjectType(value)
+			referentTypeFound = true
+		}
+	}
+	if !referentFound {
+		return nil, fmt.Errorf("no object found in tag %s", oid)
+	}
+	if !referentTypeFound {
+		return nil, fmt.Errorf("no type found in tag %s", oid)
+	}
+	return &Tag{
+		Size:         counts.NewCount32(uint64(len(data))),
+		Referent:     referent,
+		ReferentType: referentType,
+	}, nil
+}
